Article indexes as u64 instead of u32 (#25)
The RFC specifies that article counts fit in a signed int, but the limit
has been surpassed in some newsgroups, and servers are returning numbers
that would fit in a u64.

+++ examples/overviews.rs
@@ -33,9 +33,9 @@ struct Opt {
 enum Cmd {
     Xover {
         #[structopt(short, long)]
-        low: u32,
+        low: u64,
         #[structopt(short, long)]
-        high: u32,
+        high: u64,
         #[structopt(short, long, parse(from_os_str))]
         out: Option<PathBuf>,
     },
